Split HealthController into Health::TopicsController + Health::PartitionsController
Replace the single HealthController (and its LENSES/REPORT_LENSES constant
dispatch) with two explicit, namespaced controllers:

- Health::TopicsController: the aggregated top-level views - #index
  (/health/topics) and #cluster_lags (/health/cluster_lags).
- Health::PartitionsController: the per-topic drill-down, one explicit
  action per lens - #overview/#lags/#offsets/#changes (from consumer
  reports) and #cluster_lags (from the cluster).

Views are reorganised to match the derived scopes: health/topics/* and
health/partitions/*, each with its own _breadcrumbs. Routes build the two
controllers and dispatch to explicit actions.

Also:
- Order consumer groups and topics alphabetically in the aggregated
  cluster lags view, consistent with the topics view.
- Split the controller specs into topics_controller_test/
  partitions_controller_test and refresh the test timing data.
