+++ lib/karafka/web/pro/ui/routes/health.rb
@@ -37,23 +37,39 @@ module Karafka
           class Health < Base
             route do |r|
               r.on "health" do
-                controller = build(Controllers::HealthController)
+                topics_controller = build(Controllers::Health::TopicsController)
 
                 # Cluster lags is the second top-level view: a cluster-wide, routing-based lag
                 # report, aggregated per topic. Its per-partition drill-down is the `cluster_lags`
                 # lens of the per-topic view below.
                 r.get "cluster_lags" do
-                  controller.cluster_lags
+                  topics_controller.cluster_lags
                 end
 
                 # Per-topic drill-down. The per-partition lenses (overview/lags/offsets/changes and
                 # cluster_lags) live here, scoped to a single topic, instead of the old top-level
                 # all-topics views which became unusable with many topics/partitions.
                 r.on "topics", String, String do |consumer_group_id, topic_name|
-                  Controllers::HealthController::LENSES.each do |lens|
-                    r.get lens.to_s do
-                      controller.topic(consumer_group_id, topic_name, lens)
-                    end
+                  partitions_controller = build(Controllers::Health::PartitionsController)
+
+                  r.get "overview" do
+                    partitions_controller.overview(consumer_group_id, topic_name)
+                  end
+
+                  r.get "lags" do
+                    partitions_controller.lags(consumer_group_id, topic_name)
+                  end
+
+                  r.get "offsets" do
+                    partitions_controller.offsets(consumer_group_id, topic_name)
+                  end
+
+                  r.get "changes" do
+                    partitions_controller.changes(consumer_group_id, topic_name)
+                  end
+
+                  r.get "cluster_lags" do
+                    partitions_controller.cluster_lags(consumer_group_id, topic_name)
                   end
 
                   # A bare topic path defaults to the overview lens
@@ -63,13 +79,13 @@ module Karafka
                 end
 
                 r.get "topics" do
-                  controller.topics
+                  topics_controller.index
                 end
 
                 # The old top-level per-partition views are gone; keep their paths working by
                 # redirecting to the aggregated topics view so existing links/bookmarks do not break.
-                Controllers::HealthController::REPORT_LENSES.each do |lens|
-                  r.get lens.to_s do
+                %w[overview lags offsets changes].each do |legacy_lens|
+                  r.get legacy_lens do
                     r.redirect root_path("health/topics")
                   end
                 end
