+++ lib/karafka/web/pro/ui/views/health/topics/cluster_lags.erb
