+++ lib/karafka/web/pro/ui/views/health/topics/index.erb
