+++ test/timings/regular.json
@@ -168,5 +168,7 @@
   "test/lib/karafka/web/tracking/producers/listeners/booting_test.rb": 0.001,
   "test/lib/karafka/web/cli/uninstall_test.rb": 0.001,
   "test/lib/karafka/web/tracking/helpers/ttls/windows_test.rb": 0.0,
-  "test/lib/karafka/web/cli/help_test.rb": 0.0
+  "test/lib/karafka/web/cli/help_test.rb": 0.0,
+  "test/lib/karafka/web/ui/models/health/aggregated_topic_test.rb": 3.0,
+  "test/lib/karafka/web/ui/models/health/aggregated_cluster_topic_test.rb": 3.0
 }
